Move auth builders to library
