+++ app/frontend-webapp/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,5 @@
 import NextAuth from "next-auth";
-import { options } from "@/auth/AuthOptions";
+import { options } from "@/lib/auth/AuthOptions";
 
 const handler = NextAuth(options);
